[v0.88][maintainability] Split chronosense.rs by responsibility (Closes #1824)
